backend/requirements.txt
fastapi
uvicorn
pymongo
motor
openai
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
python-dotenv
pydantic
pydantic-settings
email-validator
pdfplumber

